Got the project working in IntelliJ Idea. Some gradle/dependency and such changes.

+++ core/src/com/stabilise/util/box/F32Box.java
@@ -94,7 +94,7 @@ public class F32Box implements ITag {
     @Override public boolean getAsBoolean() { return value != 0;            }
     @Override public long    getAsLong()    { return (long) value;          }
     @Override public float   getAsFloat()   { return value;                 }
-    @Override public double  getAsDouble()  { return (double) value;        }
+    @Override public double  getAsDouble()  { return value;                 }
     @Override public String  getAsString()  { return Float.toString(value); }
     
 }
