+++ core/src/com/stabilise/util/box/I32Box.java
@@ -94,8 +94,8 @@ public class I32Box implements ITag {
     
     @Override public boolean getAsBoolean() { return value != 0;              }
     @Override public int     getAsInt()     { return value;                   }
-    @Override public long    getAsLong()    { return (long) value;            }
-    @Override public double  getAsDouble()  { return (double) value;          }
+    @Override public long    getAsLong()    { return value;                   }
+    @Override public double  getAsDouble()  { return value;                   }
     @Override public String  getAsString()  { return Integer.toString(value); }
     
 }
